src/requirements.txt
keras_segmentation==0.3.0
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.6.0.66
p_tqdm==1.3.3
pandas==1.4.1
Pillow==9.2.0
scikit_image==0.19.3
skimage==0.0
tqdm==4.62.3
